requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiohttp-cors==0.7.0
aiokafka==0.8.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
confluent-kafka==2.5.0
croniter==2.0.7
Faker==26.1.0
faust-streaming==0.10.0
frozenlist==1.4.1
idna==3.7
intervaltree==3.1.0
kafka-python==2.0.2
mode-streaming==0.3.5
multidict==6.0.5
mypy-extensions==1.0.0
opentracing==2.4.0
packaging==24.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sortedcontainers==2.4.0
terminaltables==3.1.10
urllib3==2.2.2
venusian==3.0.0
yarl==1.9.4
